Added support for undirected edges which fixes issue 14. Specifically, I added two functions, getAdjacentEdgeList() and getNeighborList(), which are wrappers to call the original CyNetwork functions with an additional call to include undirected edges/neighbors.
I also updated the inital pre-processing of the graph which selects a single edge and removes the rest in the case of multigraphs.
If nodes A and B have a directed edge A->B and an undirected edge A-B, I leave both because we would lose the B->A connection otherwise.

I also updated the version of the jar file to 1.3